% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxOR.R
\name{createData}
\alias{createData}
\title{Generate simulation data}
\usage{
createData(n, seed = 123589)
}
\arguments{
\item{n}{The total number of observations to generate}

\item{seed}{Randomized seeds for ensuring reproducible results}
}
\value{
A data frame containing age and a binary outcome variable
}
\description{
Generate simulation data
}
\examples{
dataC <- createData(200)
}
